requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai-whisper==20231117
torch==2.1.1
python-multipart==0.0.6
numpy==1.24.3
ffmpeg-python==0.2.0 